+++ swagger_parser/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.2
+- Fix problem with attributes named "default" (dart keywork incompatibility)
+
 ## 0.3.1
 - Fix the problem with APIs with @
 - Fix the problem with "number" type to map "double"